@@ -14,8 +14,8 @@
         }
     ],
     "require": {
-        "php": ">=7.4",
-        "becklyn/utilities": "^1.0",
+        "php": ">=8.0",
+        "becklyn/utilities": "^2.0",
         "ramsey/uuid": "^4.0",
         "illuminate/collections": "^8.12",
         "webmozart/assert": "^1.9"
